moved ad-status, ad-type and buy-ad-payment-method-list components into their own folders
