+++ packages/p2p/src/pages/my-ads/create-ad-form-payment-methods.jsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react-lite';
 import { useStores } from 'Stores';
 import PaymentMethodCard from 'Components/my-profile/payment-methods/payment-method-card';
 import { localize } from 'Components/i18next';
-import BuyAdPaymentMethodsList from './buy-ad-payment-methods-list.jsx';
+import BuyAdPaymentMethodsList from './buy-ad-payment-methods-list/buy-ad-payment-methods-list.jsx';
 import SellAdPaymentMethodsList from './sell-ad-payment-methods-list.jsx';
 
 const CreateAdFormPaymentMethods = ({ is_sell_advert, onSelectPaymentMethods }) => {
